// frontend/src/client/types.gen.ts
export type RoleRead = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    /**
     * Permission Tags
     */
    permission_tags?: Array<string>;
    /**
     * Created At
     */
    created_at: string;
};